Separate camera script from movement script
