+++ Assets/Scripts/NetworkPlayer/NetworkPlayerCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce47de149359b074692916a955762e20
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
